+++ plugin-maven/CHANGELOG.md
@@ -4,7 +4,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
 ### Added
-- `<welcome><openUrlOnStartup>...` for opening a browser on startup. ([#65](https://github.com/equodev/equo-ide/pull/65))
+- `<welcome><openUrl>...` for opening a browser on startup. ([#65](https://github.com/equodev/equo-ide/pull/65))
 
 ## [0.12.1] - 2023-01-25
 ### Fixed